refactor(executor): remove redundant code
